// client/college-sports-plus/src/app/shared/models/roster/football-roster.model.ts
import { LeagueAthleteModel } from '../league-athlete.model';

export class FootballRosterModel {
  FirstTeamQB: LeagueAthleteModel | undefined;
  FirstTeamRB1: LeagueAthleteModel | undefined;
  FirstTeamRB2: LeagueAthleteModel | undefined;
  FirstTeamWR1: LeagueAthleteModel | undefined;
  FirstTeamWR2: LeagueAthleteModel | undefined;
  FirstTeamTE: LeagueAthleteModel | undefined;
  FirstTeamFLEX: LeagueAthleteModel | undefined;
  FirstTeamDST: LeagueAthleteModel | undefined;
  FirstTeamK: LeagueAthleteModel | undefined;
  SecondTeamQB: LeagueAthleteModel | undefined;
  SecondTeamRB1: LeagueAthleteModel | undefined;
  SecondTeamRB2: LeagueAthleteModel | undefined;
  SecondTeamWR1: LeagueAthleteModel | undefined;
  SecondTeamWR2: LeagueAthleteModel | undefined;
  SecondTeamTE: LeagueAthleteModel | undefined;
  SecondTeamFLEX: LeagueAthleteModel | undefined;
  SecondTeamDST: LeagueAthleteModel | undefined;
  SecondTeamK: LeagueAthleteModel | undefined;
  Bench1: LeagueAthleteModel | undefined;
  Bench2: LeagueAthleteModel | undefined;
  Bench3: LeagueAthleteModel | undefined;
  Bench4: LeagueAthleteModel | undefined;
  Bench5: LeagueAthleteModel | undefined;
  Bench6: LeagueAthleteModel | undefined;
  IR: LeagueAthleteModel | undefined;

  constructor() {
    this.FirstTeamQB = undefined;
    this.FirstTeamRB1 = undefined;
    this.FirstTeamRB2 = undefined;
    this.FirstTeamWR1 = undefined;
    this.FirstTeamWR2 = undefined;
    this.FirstTeamTE = undefined;
    this.FirstTeamFLEX = undefined;
    this.FirstTeamDST = undefined;
    this.FirstTeamK = undefined;
    this.SecondTeamQB = undefined;
    this.SecondTeamRB1 = undefined;
    this.SecondTeamRB2 = undefined;
    this.SecondTeamWR1 = undefined;
    this.SecondTeamWR2 = undefined;
    this.SecondTeamTE = undefined;
    this.SecondTeamFLEX = undefined;
    this.SecondTeamDST = undefined;
    this.SecondTeamK = undefined;
    this.Bench1 = undefined;
    this.Bench2 = undefined;
    this.Bench3 = undefined;
    this.Bench4 = undefined;
    this.Bench5 = undefined;
    this.Bench6 = undefined;
    this.IR = undefined;
  }
}